Belated change re merge of #2408 (Top-level Await)
The expectation for fail/1aefe47e20eb91fa.module.js
shifted from early error to parsing error.

Formerly, we had:
    ModuleItem : StatementListItem[~Yield, ~Await, ~Return]
which allowed this test's `await` to be parsed as an IdentifierReference,
but then disallowed by an early error rule.

TLA changed that ~Await to +Await,
so this test's `await` can't be parsed as an IdentifierReference,
and it can't be parsed as an AwaitExpression,
because it's missing the UnaryExpression.
So it's now a parsing error.

So the test's placement in 'fail' was formerly incorrect,
but now it's correct.