Drop extract_info_from_standard_preamble()
Formerly, each run would call this function hundreds of times.
When PR #545 was merged, that went down to a handful of times.
With recent changes in Section.py, it's now called only once,
and that just results in "Starts like std preamble but isn't".

If a standardized-wording preamble does arise (e.g., from a longstanding PR),
_handle_header_with_std_preamble will deal with it adequately.
(One thing that function doesn't do is extract parameter-types from the preamble.
That ability could be transferred over, but it seems unlikely to be used much.)