%ThrowTypeError%: Change section_kind & species
Formerly it was "anonymous_built_in_function" (which is true),
which maps to "bif: * per realm" (which is false).

Now it's "CallConstruct", which maps to "bif: intrinsic" (both true).

(I.e., "anonymous_built_in_function" is now reserved for cases in which
the mapping to "bif: * per realm" is valid.)