Eliminate the concept of 'abstract' numeric methods
Before the merge of PR #2546,
the spec had expressions like `_T_::unaryMinus(_x_)`
which suggested the idea that `::unaryMinus` is an "abstract method",
and at each execution of that expression,
it will dispatch/delegate to one of the "concrete methods"
Number::unaryMinus or BigInt::unaryMinus,
depending on the value of _T_ at the time.

But PR #2546 removed such expressions;
all references to numeric methods are directly to concrete methods.
So there's no longer any need for
the concept of abstract methods `::unaryMinus` etc.

Practically, self.alg_info_['op'] no longer has an entry for "::unaryMinus" etc,
it has entries for each of "Number::unaryMinus" and "BigInt::unaryMinus" etc.