Handle Completion Records more-or-less properly
For a long time, the ECMAScript spec would implicitly inter-convert between
`a normal completion whose [[Value]] is X` and just `X`.
So, for the purpose of static analysis,
it made sense to treat those as the same thing.

However, over the course of a few PRs
(culminating in 2547?, merged 2022-02-24)
"completion reform" eliminated almost all such implicit conversions.
(And the remaining ones are fairly clearly defined.)
An improper implicit conversion would now be an editorial error.
So in static analysis, it is now desirable to distingish those two things.