PR #3290 has been merged
Rather than adding -1 to {VAL_DESC},
it seems like it should be in {LITERAL}
(or more specifically, {MATH_LITERAL}),
but that would create an ambiguity.
{EX} could derive "-1" in two different ways:
- {EX} -> {LITERAL} -> -1
- {EX} -> {NUM_EXPR} -> {PRODUCT} -> {UNARY_OPERATOR}{FACTOR} ... -1

It looks like -1 as a {VAL_DESC} will be replaced by ~not-found~
(https://github.com/tc39/ecma262/issues/3298),
so looking for a better solution isn't worth the bother.